refactor(mcp): split tools into scope groups (core/read/work/pm/crud)
Pure code move. Zero behavior change. All 47 MCP tools remain registered
with identical names, schemas, and handler bodies.

- Splits src/lib/mcp/tools.ts (24 tools) and src/lib/mcp/roadmap-tools.ts
  (23 tools) into 5 group files under src/lib/mcp/groups/.
- Extracts shared helpers into src/lib/mcp/shared.ts.
- Rewrites src/lib/mcp/server.ts with a buildServer(groups?) factory
  defaulting to ALL_GROUPS so the existing /api/mcp route is unchanged.

Group sizes: core 6, read 9, work 18, pm 5, crud 9 = 47 total.

Sets the foundation for PR 2 (new /api/mcp/pm and /api/mcp/crud routes).
Subsequent PRs (4, 5) edit individual group files instead of the
2800-line monoliths.

Per specs/mcp-surface-v2-build-plan.md PR 1.
