+++ src/lib/mcp/groups/work.ts
@@ -1,24 +1,17 @@
 /**
- * sc-mission-control MCP tools.
+ * Work MCP tools — task execution, mail, deliverables, evidence,
+ * subtask delegation, knowledge, subagent dispatch registration, and the
+ * note-lifecycle tools (mark_note_consumed / archive_note).
  *
- * All 12 tools consolidated into one file. Each tool is a thin wrapper
- * around a service-layer function (src/lib/services/*) that handles
- * authorization, DB work, and broadcasts.
- *
- * Every state-changing tool takes `agent_id` as the required first arg —
- * the Phase 0 spike confirmed OpenClaw passes no identity to MCP servers,
- * so agents self-identify on every call. The `MC_API_TOKEN` bearer is the
- * trust boundary (same as the current HTTP API).
+ * Behavior is unchanged from the legacy `tools.ts` consolidation; this is
+ * a pure relocation as part of the MCP surface refactor (PR 1).
  */
 
 import { z } from 'zod';
 import type { McpServer } from '@modelcontextprotocol/sdk/server/mcp.js';
-import type { CallToolResult } from '@modelcontextprotocol/sdk/types.js';
 
 import { queryAll, queryOne, run } from '@/lib/db';
 import { AuthzError, assertAgentActive } from '@/lib/authz/agent-task';
-import { authzErrorToToolResult, internalErrorToToolResult } from './errors';
-import { logMcpToolCall } from './debug';
 
 import { registerDeliverable } from '@/lib/services/task-deliverables';
 import { submitEvidence, ALL_EVIDENCE_GATES } from '@/lib/services/task-evidence';
@@ -33,338 +26,25 @@ import { getUnreadMail } from '@/lib/mailbox';
 import { getOpenClawClient } from '@/lib/openclaw/client';
 import { spawnDelegationSubtask } from '@/lib/convoy';
 import { internalDispatch } from '@/lib/internal-dispatch';
-import { postPmChatMessage } from '@/lib/agents/pm-dispatch';
 import { upsertSession, type ScopeType } from '@/lib/db/mc-sessions';
 import {
   archiveNote as archiveNoteDb,
-  createNote,
   getNote,
-  listNotes,
   markNoteConsumed as markNoteConsumedDb,
-  parseAttachedFiles,
   parseConsumedStages,
-  AgentNoteValidationError,
-  NOTE_BODY_MAX,
-  type AgentNote,
-  type NoteImportance,
-  type NoteKind,
 } from '@/lib/db/agent-notes';
 import { broadcast } from '@/lib/events';
-import type { Task } from '@/lib/types';
 
-// Common shape: agent_id on every state-changing tool.
-const agentIdArg = z
-  .string()
-  .min(1)
-  .describe("The calling agent's MC agent_id (see whoami)");
+import {
+  agentIdArg,
+  taskIdArg,
+  textResult,
+  trace,
+  deriveWorkspaceFromAgent,
+  noteToPayload,
+} from '../shared';
 
-const taskIdArg = z
-  .string()
-  .min(1)
-  .describe('Task UUID');
-
-/**
- * Wraps a tool handler to log, time, and catch AuthzError uniformly.
- * The handler returns a CallToolResult or its contents; AuthzError is
- * mapped to a structured tool-error result rather than propagated as a
- * JSON-RPC protocol error.
- */
-function trace<TArgs extends { agent_id?: string; task_id?: string }>(
-  toolName: string,
-  handler: (args: TArgs) => Promise<CallToolResult> | CallToolResult,
-) {
-  return async (args: TArgs): Promise<CallToolResult> => {
-    const started = Date.now();
-    try {
-      const result = await handler(args);
-      logMcpToolCall({
-        toolName,
-        agentId: args.agent_id ?? null,
-        taskId: args.task_id ?? null,
-        ok: !result.isError,
-        durationMs: Date.now() - started,
-        error: result.isError ? extractErrorMessage(result) : undefined,
-      });
-      return result;
-    } catch (err) {
-      const durationMs = Date.now() - started;
-      if (err instanceof AuthzError) {
-        logMcpToolCall({
-          toolName,
-          agentId: args.agent_id ?? null,
-          taskId: args.task_id ?? null,
-          ok: false,
-          durationMs,
-          error: `authz:${err.code}`,
-        });
-        return authzErrorToToolResult(err);
-      }
-      logMcpToolCall({
-        toolName,
-        agentId: args.agent_id ?? null,
-        taskId: args.task_id ?? null,
-        ok: false,
-        durationMs,
-        error: err instanceof Error ? err.message : String(err),
-      });
-      return internalErrorToToolResult(err);
-    }
-  };
-}
-
-function extractErrorMessage(result: CallToolResult): string {
-  const sc = result.structuredContent;
-  if (sc && typeof sc === 'object') {
-    const obj = sc as Record<string, unknown>;
-    if (typeof obj.message === 'string' && obj.message) return obj.message;
-    if (typeof obj.error === 'string' && obj.error) return obj.error;
-  }
-  const first = result.content?.[0];
-  if (first && first.type === 'text' && first.text) return first.text;
-  return 'tool returned isError';
-}
-
-function textResult(text: string, structured?: Record<string, unknown>): CallToolResult {
-  return {
-    content: [{ type: 'text', text }],
-    ...(structured ? { structuredContent: structured } : {}),
-  };
-}
-
-/**
- * Resolve the workspace_id for the calling agent. Used by tools whose
- * scope is workspace-bound (notes spine, etc.). Strict — refuses
- * gateway_agent_id on principle: callers must pass the MC UUID, which
- * matches one row unambiguously per spec §1.4 (post-migration the
- * runner is the only org-wide gateway id; non-PM agents have NULL).
- *
- * Throws AuthzError on miss so the trace wrapper produces a clean
- * error result.
- */
-function deriveWorkspaceFromAgent(agentId: string): string {
-  const row = queryOne<{ workspace_id: string }>(
-    `SELECT workspace_id FROM agents WHERE id = ? LIMIT 1`,
-    [agentId],
-  );
-  if (row) return row.workspace_id;
-  throw new AuthzError('agent_not_found', `agent ${agentId} not found`);
-}
-
-// ─── tool registrations ─────────────────────────────────────────────
-
-export function registerAllTools(server: McpServer): void {
-  // whoami ────────────────────────────────────────────────────────
-  server.registerTool(
-    'whoami',
-    {
-      title: 'Identify the calling agent',
-      description:
-        "Returns the calling agent's identity, assigned task ids, and the peer roster (gateway_id → MC agent_id). Call this once at session start to learn your own agent_id and peers.",
-      inputSchema: {
-        agent_id: z
-          .string()
-          .min(1)
-          .describe(
-            "Your agent identity — accepts either MC agent_id (UUID) or gateway_agent_id (e.g. 'mc-project-manager'). whoami is the bootstrap call, so either form works; other tools require the MC agent_id this returns.",
-          ),
-      },
-      annotations: { readOnlyHint: true, openWorldHint: false },
-    },
-    trace('whoami', async ({ agent_id }) => {
-      // Multi-workspace gateway clones (#133) make `gateway_agent_id`
-      // ambiguous on its own. Try UUID first; on miss, only fall back to
-      // gateway_agent_id when it resolves to exactly one row. Otherwise
-      // return a structured error listing the candidate workspaces so
-      // the operator can re-dispatch with the correct UUID — silently
-      // returning the first match would mint the wrong workspace_id and
-      // break every subsequent MCP call with workspace_mismatch.
-      type AgentRow = {
-        id: string;
-        name: string;
-        role: string;
-        workspace_id: string;
-        gateway_agent_id: string | null;
-        is_active: number | null;
-      };
-      let me = queryOne<AgentRow>(
-        `SELECT id, name, role, workspace_id, gateway_agent_id, is_active
-           FROM agents WHERE id = ? LIMIT 1`,
-        [agent_id],
-      );
-      if (!me) {
-        const byGateway = queryAll<AgentRow>(
-          `SELECT id, name, role, workspace_id, gateway_agent_id, is_active
-             FROM agents WHERE gateway_agent_id = ?`,
-          [agent_id],
-        );
-        if (byGateway.length === 1) {
-          me = byGateway[0];
-        } else if (byGateway.length > 1) {
-          const candidates = byGateway.map((r) => ({ id: r.id, workspace_id: r.workspace_id, name: r.name }));
-          return {
-            isError: true,
-            content: [{
-              type: 'text',
-              text: `gateway_agent_id "${agent_id}" exists in ${byGateway.length} workspaces. Pass your MC agent_id (UUID) instead — the dispatch briefing embeds it as "Your agent_id is: …".`,
-            }],
-            structuredContent: { error: 'ambiguous_gateway_id', gateway_agent_id: agent_id, candidates },
-          };
-        }
-      }
-      if (!me) {
-        return {
-          isError: true,
-          content: [{ type: 'text', text: `agent ${agent_id} not found (tried both id and gateway_agent_id)` }],
-          structuredContent: { error: 'agent_not_found', agent_id },
-        };
-      }
-
-      const tasks = queryAll<{ id: string; title: string; status: string }>(
-        `SELECT id, title, status FROM tasks
-          WHERE workspace_id = ?
-            AND (assigned_agent_id = ?
-              OR id IN (SELECT task_id FROM task_roles WHERE agent_id = ?))
-            AND status NOT IN ('done', 'cancelled')
-          ORDER BY updated_at DESC`,
-        [me.workspace_id, me.id, me.id],
-      );
-
-      const peers = queryAll<{ id: string; gateway_agent_id: string; name: string; role: string }>(
-        `SELECT id, gateway_agent_id, name, role FROM agents
-          WHERE workspace_id = ?
-            AND gateway_agent_id IS NOT NULL AND gateway_agent_id != ''
-            AND id != ?`,
-        [me.workspace_id, me.id],
-      );
-      const peerMap: Record<string, { id: string; name: string; role: string }> = {};
-      for (const p of peers) {
-        peerMap[p.gateway_agent_id] = { id: p.id, name: p.name, role: p.role };
-      }
-
-      const payload = {
-        id: me.id,
-        name: me.name,
-        role: me.role,
-        workspace_id: me.workspace_id,
-        gateway_agent_id: me.gateway_agent_id,
-        assigned_task_ids: tasks.map((t) => t.id),
-        assigned_tasks: tasks,
-        peers: peerMap,
-      };
-      return textResult(JSON.stringify(payload, null, 2), payload);
-    }),
-  );
-
-  // get_workspace_context ─────────────────────────────────────────
-  // Lets an in-flight agent re-read the workspace's "rules of the road"
-  // markdown without re-dispatch. The same blob is prepended to every
-  // task dispatch as a "## Workspace conventions" section, so this tool
-  // is mostly useful when an agent forks a subtask and wants to keep
-  // the conventions in scope, or for debugging what the operator wrote.
-  server.registerTool(
-    'get_workspace_context',
-    {
-      title: 'Get workspace conventions',
-      description:
-        "Returns the markdown 'rules of the road' the operator wrote for this workspace (repos, testing, git/PR rules, package manager, etc). Same content prepended to every task dispatch's prompt; call this when you need to re-read it mid-session or surface it to a delegated subagent.",
-      inputSchema: {
-        agent_id: z
-          .string()
-          .min(1)
-          .describe(
-            "Your MC agent_id (UUID) or gateway_agent_id. Either form works — looked up the same way as whoami.",
-          ),
-      },
-      annotations: { readOnlyHint: true, openWorldHint: false },
-    },
-    trace('get_workspace_context', async ({ agent_id }) => {
-      // Same multi-workspace ambiguity guard as whoami — picking the
-      // first row by gateway_agent_id would surface the wrong
-      // workspace's context_md silently.
-      let me = queryOne<{ workspace_id: string }>(
-        `SELECT workspace_id FROM agents WHERE id = ? LIMIT 1`,
-        [agent_id],
-      );
-      if (!me) {
-        const byGateway = queryAll<{ id: string; workspace_id: string; name: string }>(
-          `SELECT id, workspace_id, name FROM agents WHERE gateway_agent_id = ?`,
-          [agent_id],
-        );
-        if (byGateway.length === 1) {
-          me = { workspace_id: byGateway[0].workspace_id };
-        } else if (byGateway.length > 1) {
-          return {
-            isError: true,
-            content: [{
-              type: 'text',
-              text: `gateway_agent_id "${agent_id}" exists in ${byGateway.length} workspaces. Pass your MC agent_id (UUID) instead.`,
-            }],
-            structuredContent: {
-              error: 'ambiguous_gateway_id',
-              gateway_agent_id: agent_id,
-              candidates: byGateway,
-            },
-          };
-        }
-      }
-      if (!me) {
-        return {
-          isError: true,
-          content: [{ type: 'text', text: `agent ${agent_id} not found` }],
-          structuredContent: { error: 'agent_not_found', agent_id },
-        };
-      }
-      const ws = queryOne<{ id: string; name: string; context_md: string | null }>(
-        `SELECT id, name, context_md FROM workspaces WHERE id = ?`,
-        [me.workspace_id],
-      );
-      const payload = {
-        workspace_id: me.workspace_id,
-        workspace_name: ws?.name ?? null,
-        context_md: ws?.context_md ?? null,
-        present: typeof ws?.context_md === 'string' && ws.context_md.trim().length > 0,
-      };
-      return textResult(JSON.stringify(payload, null, 2), payload);
-    }),
-  );
-
-  // list_peers ────────────────────────────────────────────────────
-  server.registerTool(
-    'list_peers',
-    {
-      title: 'List peer agents in your workspace',
-      description:
-        'Returns the workspace peer roster (gateway_id → MC agent_id, name, role). Use this before send_mail or delegate when you need a recipient id.',
-      inputSchema: { agent_id: agentIdArg },
-      annotations: { readOnlyHint: true, openWorldHint: false },
-    },
-    trace('list_peers', async ({ agent_id }) => {
-      const me = queryOne<{ workspace_id: string }>(
-        `SELECT workspace_id FROM agents WHERE id = ? LIMIT 1`,
-        [agent_id],
-      );
-      if (!me) {
-        return {
-          isError: true,
-          content: [{ type: 'text', text: `agent ${agent_id} not found` }],
-          structuredContent: { error: 'agent_not_found', agent_id },
-        };
-      }
-      const peers = queryAll<{
-        id: string;
-        gateway_agent_id: string | null;
-        name: string;
-        role: string;
-      }>(
-        `SELECT id, gateway_agent_id, name, role FROM agents
-          WHERE workspace_id = ? AND id != ?
-          ORDER BY role, name`,
-        [me.workspace_id, agent_id],
-      );
-      return textResult(JSON.stringify({ peers }, null, 2), { peers });
-    }),
-  );
-
+export function registerWorkTools(server: McpServer): void {
   // get_task ──────────────────────────────────────────────────────
   server.registerTool(
     'get_task',
@@ -590,239 +270,6 @@ export function registerAllTools(server: McpServer): void {
         reject_reason: result.rejectReason,
       };
       return textResult(JSON.stringify(summary, null, 2), summary);
-    }),
-  );
-
-  // log_activity ──────────────────────────────────────────────────
-  server.registerTool(
-    'log_activity',
-    {
-      title: 'Log an activity against a task',
-      description:
-        "Append a progress note. Required before status transitions — the evidence gate rejects transitions without at least one activity. Use activity_type='completed' for your final message before transitioning.",
-      inputSchema: {
-        agent_id: agentIdArg,
-        task_id: taskIdArg,
-        activity_type: z.enum([
-          'spawned',
-          'updated',
-          'completed',
-          'file_created',
-          'status_changed',
-        ]),
-        message: z.string().min(1).max(5000),
-        metadata: z.record(z.string(), z.unknown()).optional(),
-      },
-      annotations: { destructiveHint: true, openWorldHint: false },
-    },
-    trace('log_activity', async (args) => {
-      const activity = logActivity({
-        taskId: args.task_id,
-        actingAgentId: args.agent_id,
-        activityType: args.activity_type,
-        message: args.message,
-        metadata: args.metadata ? JSON.stringify(args.metadata) : undefined,
-      });
-      // Read post-write counts back. The "evidence" count matches what
-      // the evidence gate inspects (completed / file_created / updated)
-      // so the agent can see whether it's met the bar without guessing.
-      const totals = queryOne<{ total_count: number; evidence_count: number }>(
-        `SELECT
-           COUNT(*) as total_count,
-           SUM(CASE WHEN activity_type IN ('completed','file_created','updated') THEN 1 ELSE 0 END) as evidence_count
-         FROM task_activities WHERE task_id = ?`,
-        [args.task_id],
-      );
-      const summary = {
-        activity,
-        total_activities_on_task: Number(totals?.total_count ?? 0),
-        evidence_qualifying_activities_on_task: Number(totals?.evidence_count ?? 0),
-      };
-      return textResult(JSON.stringify(summary, null, 2), summary);
-    }),
-  );
-
-  // ── Notes spine (scope-keyed sessions Phase A) ───────────────────
-  // take_note / read_notes / mark_note_consumed / archive_note
-  // See specs/scope-keyed-sessions.md §3 for the full design.
-  // No agents call these yet — this is the surface area; role-souls
-  // start using it in Phase C when the notetaker addendum lands.
-
-  const noteKindArg = z
-    .enum(['discovery', 'blocker', 'uncertainty', 'decision', 'observation', 'question', 'breadcrumb'])
-    .describe('What kind of note this is. See agent-templates/_shared/notetaker.md for guidance.');
-
-  const noteImportanceArg = z
-    .union([z.literal(0), z.literal(1), z.literal(2)])
-    .describe('0 = low (default), 1 = normal, 2 = high (PM Chat surfaces this in real time).');
-
-  function noteToPayload(note: AgentNote): Record<string, unknown> {
-    return {
-      id: note.id,
-      workspace_id: note.workspace_id,
-      agent_id: note.agent_id,
-      task_id: note.task_id,
-      initiative_id: note.initiative_id,
-      scope_key: note.scope_key,
-      role: note.role,
-      run_group_id: note.run_group_id,
-      kind: note.kind,
-      audience: note.audience,
-      body: note.body,
-      attached_files: parseAttachedFiles(note),
-      importance: note.importance,
-      created_at: note.created_at,
-    };
-  }
-
-  // take_note ──────────────────────────────────────────────────────
-  server.registerTool(
-    'take_note',
-    {
-      title: 'Record an observation, decision, blocker, or breadcrumb',
-      description:
-        "Cheap, spammable observability primitive. Use liberally — every meaningful moment of your work should leave a trail here. NOT evidence: notes don't unblock status transitions (use log_activity for that). Set importance=2 only for genuinely high-stakes findings (security issues, broken assumptions); the PM sees those in PM Chat in real time.",
-      inputSchema: {
-        agent_id: agentIdArg,
-        kind: noteKindArg,
-        body: z
-          .string()
-          .min(1)
-          .max(NOTE_BODY_MAX)
-          .describe('Concrete > aspirational. One thought per note. Reference file paths in attached_files.'),
-        scope_key: z
-          .string()
-          .min(1)
-          .describe('The openclaw sessionKey you are running under. Take this verbatim from your dispatch briefing.'),
-        role: z
-          .string()
-          .min(1)
-          .describe("Your role-of-the-moment ('builder', 'tester', 'pm', 'researcher', etc.). From the briefing."),
-        run_group_id: z
-          .string()
-          .min(1)
-          .describe('UUID minted at session start that groups all notes from one run/stage. From the briefing.'),
-        task_id: z.string().optional().describe('Set when the note relates to a specific task.'),
-        initiative_id: z
-          .string()
-          .optional()
-          .describe('Set when the note relates to an initiative directly (no task scope).'),
-        audience: z
-          .string()
-          .optional()
-          .describe("Who this note is for: 'pm', 'reviewer', 'next-stage', 'tester', etc. Omit for anyone."),
-        attached_files: z
-          .array(z.string())
-          .optional()
-          .describe('File paths the note references. Helps the next session navigate without re-reading.'),
-        importance: noteImportanceArg.optional(),
-      },
-      annotations: { destructiveHint: false, openWorldHint: false },
-    },
-    trace('take_note', async (args) => {
-      try {
-        const note = createNote({
-          workspace_id: deriveWorkspaceFromAgent(args.agent_id),
-          agent_id: args.agent_id,
-          task_id: args.task_id ?? null,
-          initiative_id: args.initiative_id ?? null,
-          scope_key: args.scope_key,
-          role: args.role,
-          run_group_id: args.run_group_id,
-          kind: args.kind as NoteKind,
-          audience: args.audience ?? null,
-          body: args.body,
-          attached_files: args.attached_files ?? null,
-          importance: (args.importance ?? 0) as NoteImportance,
-        });
-
-        const payload = noteToPayload(note);
-        broadcast({ type: 'agent_note_created', payload });
-
-        // D5: importance=2 notes auto-post to PM Chat so the operator
-        // sees high-stakes findings in their primary chat surface in
-        // real time. Best-effort: failures don't fail the take_note.
-        if (note.importance === 2) {
-          try {
-            const fileTrail = parseAttachedFiles(note);
-            const filesLine = fileTrail.length > 0
-              ? `\n\nFiles: ${fileTrail.map((f) => `\`${f}\``).join(', ')}`
-              : '';
-            postPmChatMessage({
-              workspace_id: note.workspace_id,
-              role: 'assistant',
-              content:
-                `**🚩 ${note.kind} (from ${note.role})**\n\n${note.body}${filesLine}`,
-            });
-          } catch (chatErr) {
-            console.warn(
-              '[take_note] importance=2 PM Chat post failed:',
-              (chatErr as Error).message,
-            );
-          }
-        }
-
-        return textResult(JSON.stringify(payload, null, 2), payload);
-      } catch (err) {
-        if (err instanceof AgentNoteValidationError) {
-          return {
-            isError: true,
-            content: [{ type: 'text', text: err.message }],
-            structuredContent: { error: 'validation', message: err.message },
-          };
-        }
-        throw err;
-      }
-    }),
-  );
-
-  // read_notes ─────────────────────────────────────────────────────
-  server.registerTool(
-    'read_notes',
-    {
-      title: 'List notes visible to the calling agent',
-      description:
-        "Query notes — by task, by initiative, by audience, by kind. Use this BEFORE committing to an approach: scan for prior decisions, blockers, and breadcrumbs from earlier stages. Returns up to 50 by default (capped at 200). Default order is created_at ASC so prior context comes first.",
-      inputSchema: {
-        agent_id: agentIdArg,
-        task_id: z.string().optional(),
-        initiative_id: z.string().optional(),
-        audience: z
-          .string()
-          .optional()
-          .describe("Restrict to notes addressed to this audience or to anyone (NULL audience). Common values: 'pm', 'next-stage', 'reviewer', or your own role."),
-        kinds: z.array(noteKindArg).optional().describe('Restrict to specific kinds.'),
-        not_consumed_by_stage: z
-          .string()
-          .optional()
-          .describe('Skip notes already marked consumed by this stage slug. Useful when filtering "what is new for me".'),
-        scope_key: z.string().optional(),
-        run_group_id: z.string().optional(),
-        min_importance: z.union([z.literal(0), z.literal(1), z.literal(2)]).optional(),
-        include_archived: z.boolean().optional(),
-        limit: z.number().int().positive().max(200).optional(),
-        order: z.enum(['asc', 'desc']).optional(),
-      },
-      annotations: { readOnlyHint: true, openWorldHint: false },
-    },
-    trace('read_notes', async (args) => {
-      const workspaceId = deriveWorkspaceFromAgent(args.agent_id);
-      const notes = listNotes({
-        workspace_id: workspaceId,
-        task_id: args.task_id,
-        initiative_id: args.initiative_id,
-        audience: args.audience,
-        kinds: args.kinds as ReadonlyArray<NoteKind> | undefined,
-        not_consumed_by_stage: args.not_consumed_by_stage,
-        scope_key: args.scope_key,
-        run_group_id: args.run_group_id,
-        min_importance: args.min_importance as NoteImportance | undefined,
-        include_archived: args.include_archived,
-        limit: args.limit,
-        order: args.order,
-      });
-      const payload = { count: notes.length, notes: notes.map(noteToPayload) };
-      return textResult(JSON.stringify(payload, null, 2), payload);
     }),
   );
